# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ServicesHub B2B intermediation platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /provider/dashboard:
    get:
      operationId: getProviderDashboard
      tags: [provider]
      summary: Provider dashboard summary
      parameters:
        - $ref: "#/components/parameters/ProfileHeader"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderDashboard"
  /provider/catalog:
    get:
      operationId: listProviderCatalog
      tags: [provider]
      summary: List provider's own catalog items
      parameters:
        - $ref: "#/components/parameters/ProfileHeader"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CatalogItem"
    post:
      operationId: createCatalogItem
      tags: [provider]
      summary: Create a new catalog item
      parameters:
        - $ref: "#/components/parameters/ProfileHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCatalogItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatalogItem"
  /provider/catalog/{id}:
    put:
      operationId: updateCatalogItem
      tags: [provider]
      summary: Update a catalog item
      parameters:
        - $ref: "#/components/parameters/ProfileHeader"
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCatalogItemInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatalogItem"
    delete:
      operationId: deleteCatalogItem
      tags: [provider]
      summary: Delete a catalog item
      parameters:
        - $ref: "#/components/parameters/ProfileHeader"
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No content
  /provider/work-orders:
    get:
      operationId: listProviderWorkOrders
      tags: [provider]
      summary: List work orders assigned to provider
      parameters:
        - $ref: "#/components/parameters/ProfileHeader"
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkOrderProviderView"
  /provider/work-orders/{id}:
    get:
      operationId: getProviderWorkOrder
      tags: [provider]
      summary: Get single work order for provider
      parameters:
        - $ref: "#/components/parameters/ProfileHeader"
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkOrderProviderView"
  /provider/work-orders/{id}/action:
    post:
      operationId: providerWorkOrderAction
      tags: [provider]
      summary: Take action on a work order (accept, start, complete)
      parameters:
        - $ref: "#/components/parameters/ProfileHeader"
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkOrderActionInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkOrderProviderView"
  /provider/financial:
    get:
      operationId: getProviderFinancial
      tags: [provider]
      summary: Provider financial summary and payout history
      parameters:
        - $ref: "#/components/parameters/ProfileHeader"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderFinancial"

  # ─── Admin Portal ────────────────────────────────────────────────
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications for current profile
      parameters:
        - $ref: "#/components/parameters/ProfileHeader"
        - name: unreadOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - $ref: "#/components/parameters/ProfileHeader"
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
  /notifications/read-all:
    post:
      operationId: markAllNotificationsRead
      tags: [notifications]
      summary: Mark all notifications as read
      parameters:
        - $ref: "#/components/parameters/ProfileHeader"
      responses:
        "204":
          description: No content
components:
  parameters:
    ProfileHeader:
      name: x-profile-id
      in: header
      required: false
      schema:
        type: string
  schemas:
...
    CatalogItem:
      type: object
      properties:
        id:
          type: string
        providerCompanyId:
          type: string
        providerCompanyName:
          type: string
        name:
          type: string
        description:
          type: string
        category:
          type: string
        estimatedDays:
          type: integer
        basePrice:
          type: number
        isAvailable:
          type: boolean
        regions:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required: [id, providerCompanyId, name, category, basePrice, isAvailable, createdAt]
...
    CreateCatalogItemInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
        estimatedDays:
          type: integer
        basePrice:
          type: number
        isAvailable:
          type: boolean
        regions:
          type: array
          items:
            type: string
      required: [name, category, basePrice]

    # ─── Work Orders ──────────────────────────────────────────────
    WorkOrderStatus:
      type: string
      enum:
        - requested
        - accepted
        - in_progress
        - completed
        - invoiced
        - paid
        - paid_out
        - closed
        - cancelled
...
    WorkOrderProviderView:
      type: object
      properties:
        id:
          type: string
        serviceName:
          type: string
        category:
          type: string
        location:
          type: string
        description:
          type: string
        notes:
          type: string
        status:
          $ref: "#/components/schemas/WorkOrderStatus"
        requestedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        basePrice:
          type: number
        travelCost:
          type: number
        providerReceivable:
          type: number
        statusHistory:
          type: array
          items:
            $ref: "#/components/schemas/StatusHistoryEntry"
      required: [id, serviceName, status, requestedAt]
...
    StatusHistoryEntry:
      type: object
      properties:
        id:
          type: string
        status:
          $ref: "#/components/schemas/WorkOrderStatus"
        note:
          type: string
        changedAt:
          type: string
          format: date-time
        changedBy:
          type: string
      required: [id, status, changedAt]
...
    WorkOrderActionInput:
      type: object
      properties:
        action:
          type: string
          enum: [accept, start, complete]
        notes:
          type: string
      required: [action]
...
    Payout:
      type: object
      properties:
        id:
          type: string
        providerCompanyId:
          type: string
        providerCompanyName:
          type: string
        amount:
          type: number
        workOrderIds:
          type: array
          items:
            type: string
        notes:
          type: string
        paidAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, providerCompanyId, amount, paidAt, createdAt]
...
    Notification:
      type: object
      properties:
        id:
          type: string
        recipientId:
          type: string
        recipientRole:
          type: string
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum:
            - new_request
            - request_accepted
            - service_started
            - service_completed
            - invoice_generated
            - payment_confirmed
            - payout_confirmed
        relatedWorkOrderId:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, recipientId, title, message, type, isRead, createdAt]

    # ─── Dashboards ──────────────────────────────────────────────
...
    ProviderDashboard:
      type: object
      properties:
        newOrders:
          type: integer
        inProgress:
          type: integer
        completed:
          type: integer
        totalReceivable:
          type: number
        totalReceived:
          type: number
        recentWorkOrders:
          type: array
          items:
            $ref: "#/components/schemas/WorkOrderProviderView"
      required: [newOrders, inProgress, completed, totalReceivable, totalReceived]
...
    ProviderFinancial:
      type: object
      properties:
        totalReceivable:
          type: number
        totalReceived:
          type: number
        totalTravelCost:
          type: number
        pendingWorkOrders:
          type: array
          items:
            $ref: "#/components/schemas/WorkOrderProviderView"
        payouts:
          type: array
          items:
            $ref: "#/components/schemas/Payout"
      required: [totalReceivable, totalReceived, totalTravelCost]