# lib/api-zod/src/generated/types/providerDashboard.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ServicesHub B2B intermediation platform API
 * OpenAPI spec version: 0.1.0
 */
import type { WorkOrderProviderView } from "./workOrderProviderView";

export interface ProviderDashboard {
  newOrders: number;
  inProgress: number;
  completed: number;
  totalReceivable: number;
  totalReceived: number;
  recentWorkOrders?: WorkOrderProviderView[];
}