# lib/api-zod/src/generated/types/workOrderProviderView.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ServicesHub B2B intermediation platform API
 * OpenAPI spec version: 0.1.0
 */
import type { StatusHistoryEntry } from "./statusHistoryEntry";
import type { WorkOrderStatus } from "./workOrderStatus";

export interface WorkOrderProviderView {
  id: string;
  serviceName: string;
  category?: string;
  location?: string;
  description?: string;
  notes?: string;
  status: WorkOrderStatus;
  requestedAt: Date;
  completedAt?: Date;
  basePrice?: number;
  travelCost?: number;
  providerReceivable?: number;
  statusHistory?: StatusHistoryEntry[];
}