# lib/api-zod/src/generated/types/workOrderStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ServicesHub B2B intermediation platform API
 * OpenAPI spec version: 0.1.0
 */

export type WorkOrderStatus =
  (typeof WorkOrderStatus)[keyof typeof WorkOrderStatus];

export const WorkOrderStatus = {
  requested: "requested",
  accepted: "accepted",
  in_progress: "in_progress",
  completed: "completed",
  invoiced: "invoiced",
  paid: "paid",
  paid_out: "paid_out",
  closed: "closed",
  cancelled: "cancelled",
} as const;